+++ spiffworkflow-backend/src/spiffworkflow_backend/config/unit_testing.py
@@ -17,6 +17,9 @@ SPIFFWORKFLOW_BACKEND_AUTH_CONFIGS = None
 SPIFFWORKFLOW_BACKEND_LOG_LEVEL = environ.get("SPIFFWORKFLOW_BACKEND_LOG_LEVEL", default="debug")
 SPIFFWORKFLOW_BACKEND_GIT_COMMIT_ON_SAVE = False
 
+SPIFFWORKFLOW_BACKEND_WEBHOOK_PROCESS_MODEL_IDENTIFIER = "test_group/simple_script"
+SPIFFWORKFLOW_BACKEND_GITHUB_WEBHOOK_SECRET = "test_github_webhook_secret"
+
 # NOTE: set this here since nox shoves tests and src code to
 # different places and this allows us to know exactly where we are at the start
 worker_id = environ.get("PYTEST_XDIST_WORKER")
