+++ spiffworkflow-backend/src/spiffworkflow_backend/routes/process_api_blueprint.py
@@ -210,48 +210,6 @@ def process_data_file_download(
     )
 
 
-# sample body:
-# {"ref": "refs/heads/main", "repository": {"name": "sample-process-models",
-# "full_name": "sartography/sample-process-models", "private": False .... }}
-# test with: ngrok http 7000
-# or with:
-# npm install -g localtunnel && lt --port 7000 --subdomain oh-so-hot
-# where 7000 is the port the app is running on locally
-def github_webhook_receive(body: dict) -> Response:
-    _enforce_github_auth()
-    result = GitService.handle_web_hook(body)
-    return Response(json.dumps({"git_pull": result}), status=200, mimetype="application/json")
-
-
-def webhook(body: dict) -> Response:
-    if current_app.config["SPIFFWORKFLOW_BACKEND_WEBHOOK_ENFORCES_GITHUB_AUTH"] is True:
-        _enforce_github_auth()
-
-    if current_app.config["SPIFFWORKFLOW_BACKEND_WEBHOOK_PROCESS_MODEL_IDENTIFIER"] is None:
-        error_message = "Webhook process model implementation not configured"
-        raise ApiError(
-            error_code="webhook_not_configured",
-            message=error_message,
-            status_code=501,
-        )
-
-    process_model = _get_process_model_for_instantiation(
-        _un_modify_modified_process_model_id(current_app.config["SPIFFWORKFLOW_BACKEND_WEBHOOK_PROCESS_MODEL_IDENTIFIER"])
-    )
-    ProcessInstanceService.create_and_run_process_instance(
-        process_model=process_model,
-        persistence_level="none",
-        data_to_inject={"headers": request.headers, "body": body},
-    )
-
-    return Response(json.dumps({"ok": True}), status=200, mimetype="application/json")
-
-
-def _enforce_github_auth() -> None:
-    auth_header = request.headers.get("X-Hub-Signature-256")
-    AuthenticationService.verify_sha256_token(auth_header)
-
-
 def _get_required_parameter_or_raise(parameter: str, post_body: dict[str, Any]) -> Any:
     return_value = None
     if parameter in post_body:
