+++ spiffworkflow-backend/src/spiffworkflow_backend/services/process_instance_service.py
@@ -574,6 +574,7 @@ class ProcessInstanceService:
         cls,
         process_model: ProcessModelInfo,
         persistence_level: str,
+        user: UserModel,
         data_to_inject: dict | None = None,
         process_id_to_run: str | None = None,
     ) -> ProcessInstanceProcessor:
@@ -581,14 +582,14 @@ class ProcessInstanceService:
         if persistence_level == "none":
             process_instance = ProcessInstanceModel(
                 status=ProcessInstanceStatus.not_started.value,
-                process_initiator_id=g.user.id,
+                process_initiator_id=user.id,
                 process_model_identifier=process_model.id,
                 process_model_display_name=process_model.display_name,
                 persistence_level=persistence_level,
             )
         else:
             process_instance = ProcessInstanceService.create_process_instance_from_process_model_identifier(
-                process_model.id, g.user
+                process_model.id, user
             )
 
         processor = None
